Methods return sample.id and name of outcome column
Moved return of sample.id from internal function to methods.
Data.frame method does not keep track of sample.id.
match.call is not informative inside S4 methods. Instead, set
outcome vector initially to be a 1D matrix with the column name
preserved. This is carried through to the "outcome" slot of the
null model object.