bug fix for aggregate tests with allele matching
If variants are filtered based on ref/alt alleles, the number of
variants in the genotype matrix max not correspond to the number
of variants in the GDS filter. This results in getting the wrong
length chromosome vector in the allele frequency internal function.
Add an index to pass to the .alleleFreq function to correctly
match chromosome values to variants in the geno matrix.